requirements.txt
openai
anthropic
elevenlabs
ffmpeg-python
whisper
yt-dlp
imapclient
flask
python-dotenv
